serve static files via actix_files

+++ src/routes/delete.rs
@@ -1,16 +1,16 @@
-use actix_web::{web, Error, HttpRequest, HttpResponse};
+use super::{render_unauthorized, user_can_access_page};
+use crate::prelude::*;
+use actix_session::Session;
 use actix_web::http::header;
-use actix_session::{Session};
-use crate::{prelude::*};
-use tera::{Context};
-use super::{ user_can_access_page, render_unauthorized};
+use actix_web::{web, Error, HttpRequest, HttpResponse};
+use tera::Context;
 
 pub async fn delete<T: ActixAdminAppDataTrait, E: ActixAdminViewModelTrait>(
     session: Session,
     _req: HttpRequest,
     data: web::Data<T>,
     _text: String,
-    id: web::Path<i32>
+    id: web::Path<i32>,
 ) -> Result<HttpResponse, Error> {
     let actix_admin = data.get_actix_admin();
     let entity_name = E::get_entity_name();
@@ -32,16 +32,25 @@ pub async fn delete<T: ActixAdminAppDataTrait, E: ActixAdminViewModelTrait>(
         (Ok(model), Ok(_)) => {
             for field in view_model.fields {
                 if field.field_type == ActixAdminViewModelFieldType::FileUpload {
-                    let file_name = model.get_value::<String>(&field.field_name, true, true).unwrap_or_default();
-                    let file_path = format!("{}/{}/{}", actix_admin.configuration.file_upload_directory, E::get_entity_name(), file_name.unwrap_or_default());
-                    std::fs::remove_file(file_path)?;
+                    let file_name = model
+                        .get_value::<String>(&field.field_name, true, true)
+                        .unwrap_or_default();
+                    if file_name.is_some() {
+                        let file_path = format!(
+                            "{}/{}/{}",
+                            actix_admin.configuration.file_upload_directory,
+                            E::get_entity_name(),
+                            file_name.unwrap()
+                        );
+                        std::fs::remove_file(file_path)?;
+                    }
                 }
             }
 
             Ok(HttpResponse::Ok().finish())
-        },
-        (_,_) => Ok(HttpResponse::InternalServerError().finish())
-    }   
+        }
+        (_, _) => Ok(HttpResponse::InternalServerError().finish()),
+    }
 }
 
 pub async fn delete_many<T: ActixAdminAppDataTrait, E: ActixAdminViewModelTrait>(
@@ -61,15 +70,15 @@ pub async fn delete_many<T: ActixAdminAppDataTrait, E: ActixAdminViewModelTrait>
         ctx.insert("render_partial", &true);
         return render_unauthorized(&ctx);
     }
-    
+
     let db = &data.get_db();
     let entity_name = E::get_entity_name();
     let entity_ids: Vec<i32> = text
         .split("&")
         .filter(|id| !id.is_empty())
-        .map(|id_str| id_str.replace("ids=", "").parse::<i32>().unwrap()
-    ).collect();
-    
+        .map(|id_str| id_str.replace("ids=", "").parse::<i32>().unwrap())
+        .collect();
+
     // TODO: implement delete_many
     for id in entity_ids {
         let model_result = E::get_entity(db, id).await;
@@ -79,27 +88,30 @@ pub async fn delete_many<T: ActixAdminAppDataTrait, E: ActixAdminViewModelTrait>
             (Ok(_), Ok(model)) => {
                 for field in view_model.fields {
                     if field.field_type == ActixAdminViewModelFieldType::FileUpload {
-                        let file_name = model.get_value::<String>(&field.field_name, true, true).unwrap_or_default();
-                        let file_path = format!("{}/{}/{}", actix_admin.configuration.file_upload_directory, E::get_entity_name(), file_name.unwrap_or_default());
+                        let file_name = model
+                            .get_value::<String>(&field.field_name, true, true)
+                            .unwrap_or_default();
+                        let file_path = format!(
+                            "{}/{}/{}",
+                            actix_admin.configuration.file_upload_directory,
+                            E::get_entity_name(),
+                            file_name.unwrap_or_default()
+                        );
                         std::fs::remove_file(file_path)?;
                     }
                 }
-            },
-            (Ok(_), Err(e)) => errors.push(e)
+            }
+            (Ok(_), Err(e)) => errors.push(e),
         }
     }
 
     match errors.is_empty() {
-        true => {
-            Ok(HttpResponse::SeeOther()
+        true => Ok(HttpResponse::SeeOther()
             .append_header((
                 header::LOCATION,
                 format!("/admin/{}/list?render_partial=true", entity_name),
             ))
-            .finish())
-        },
-        false => {
-            Ok(HttpResponse::InternalServerError().finish())
-        }
+            .finish()),
+        false => Ok(HttpResponse::InternalServerError().finish()),
     }
 }
